Create PaymentSheetCommonModule for shared dependencies
