+++ payments-core/src/main/java/com/stripe/android/paymentsheet/injection/FlowControllerModule.kt
@@ -3,18 +3,13 @@ package com.stripe.android.paymentsheet.injection
 import android.content.Context
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelStoreOwner
-import com.stripe.android.Logger
 import com.stripe.android.PaymentConfiguration
 import com.stripe.android.networking.AnalyticsRequestFactory
-import com.stripe.android.payments.core.injection.ENABLE_LOGGING
 import com.stripe.android.payments.core.injection.IOContext
 import com.stripe.android.paymentsheet.DefaultPrefsRepository
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PaymentSheet.FlowController
 import com.stripe.android.paymentsheet.PrefsRepository
-import com.stripe.android.paymentsheet.analytics.DefaultDeviceIdRepository
-import com.stripe.android.paymentsheet.analytics.DefaultEventReporter
-import com.stripe.android.paymentsheet.analytics.DeviceIdRepository
 import com.stripe.android.paymentsheet.analytics.EventReporter
 import com.stripe.android.paymentsheet.analytics.PaymentSheetEvent
 import com.stripe.android.paymentsheet.flowcontroller.DefaultFlowControllerInitializer
@@ -25,8 +20,6 @@ import dagger.Binds
 import dagger.Lazy
 import dagger.Module
 import dagger.Provides
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Named
 import javax.inject.Provider
 import javax.inject.Singleton
 import kotlin.coroutines.CoroutineContext
@@ -38,22 +31,7 @@ internal abstract class FlowControllerModule {
         defaultFlowControllerInitializer: DefaultFlowControllerInitializer
     ): FlowControllerInitializer
 
-    @Binds
-    abstract fun bindsEventReporter(eventReporter: DefaultEventReporter): EventReporter
-
-    @Binds
-    abstract fun bindsDeviceIdRepository(repository: DefaultDeviceIdRepository): DeviceIdRepository
-
     companion object {
-        @Provides
-        @Named(ENABLE_LOGGING)
-        fun provideEnabledLogging(): Boolean = false
-
-        @Provides
-        @Singleton
-        fun provideLogger(@Named(ENABLE_LOGGING) enableLogging: Boolean) =
-            Logger.getInstance(enableLogging)
-
         /**
          * [FlowController]'s clientSecret might be updated multiple times through
          * [FlowController.configureWithSetupIntent] or [FlowController.configureWithPaymentIntent].
@@ -92,10 +70,6 @@ internal abstract class FlowControllerModule {
             { lazyPaymentConfiguration.get().publishableKey },
             setOf(PaymentSheetEvent.PRODUCT_USAGE)
         )
-
-        @Provides
-        @IOContext
-        fun provideWorkContext(): CoroutineContext = Dispatchers.IO
 
         @Provides
         @Singleton
