+++ payments-core/src/main/java/com/stripe/android/paymentsheet/injection/PaymentSheetCommonModule.kt
@@ -1,0 +1,42 @@
+package com.stripe.android.paymentsheet.injection
+
+import com.stripe.android.BuildConfig
+import com.stripe.android.Logger
+import com.stripe.android.payments.core.injection.ENABLE_LOGGING
+import com.stripe.android.payments.core.injection.IOContext
+import com.stripe.android.paymentsheet.analytics.DefaultDeviceIdRepository
+import com.stripe.android.paymentsheet.analytics.DefaultEventReporter
+import com.stripe.android.paymentsheet.analytics.DeviceIdRepository
+import com.stripe.android.paymentsheet.analytics.EventReporter
+import dagger.Binds
+import dagger.Module
+import dagger.Provides
+import kotlinx.coroutines.Dispatchers
+import javax.inject.Named
+import javax.inject.Singleton
+import kotlin.coroutines.CoroutineContext
+
+@Module
+internal abstract class PaymentSheetCommonModule {
+
+    @Binds
+    abstract fun bindsEventReporter(eventReporter: DefaultEventReporter): EventReporter
+
+    @Binds
+    abstract fun bindsDeviceIdRepository(repository: DefaultDeviceIdRepository): DeviceIdRepository
+
+    companion object {
+        @Provides
+        @Named(ENABLE_LOGGING)
+        fun provideEnabledLogging(): Boolean = BuildConfig.DEBUG
+
+        @Provides
+        @Singleton
+        fun provideLogger(@Named(ENABLE_LOGGING) enableLogging: Boolean) =
+            Logger.getInstance(enableLogging)
+
+        @Provides
+        @IOContext
+        fun provideWorkContext(): CoroutineContext = Dispatchers.IO
+    }
+}
